Wrap up initial local implementation of queue and event store
This implementation should suffice for single-server setups that
do not need to handle large amounts of data. With Stuart Sierra's
Component framework, it should be straightforward to swap out
distributed or redis-backed structures on a "real" system. I have
not thought this through very thoroughly though, so it may be more
work than I am imagining.